sidechain commands relocation
