+++ command/validator/stake/params.go
@@ -6,7 +6,7 @@ import (
 	"math/big"
 
 	"github.com/0xPolygon/polygon-edge/command/helper"
-	validatorHelper "github.com/0xPolygon/polygon-edge/command/validator"
+	validatorHelper "github.com/0xPolygon/polygon-edge/command/validator/helper"
 )
 
 var supernetIDFlag = "supernet-id"
