+++ command/validator/withdraw/withdraw.go
@@ -12,7 +12,7 @@ import (
 	bridgeHelper "github.com/0xPolygon/polygon-edge/command/bridge/helper"
 	"github.com/0xPolygon/polygon-edge/command/helper"
 	polybftsecrets "github.com/0xPolygon/polygon-edge/command/secrets/init"
-	validatorHelper "github.com/0xPolygon/polygon-edge/command/validator"
+	validatorHelper "github.com/0xPolygon/polygon-edge/command/validator/helper"
 	"github.com/0xPolygon/polygon-edge/consensus/polybft/contractsapi"
 	"github.com/0xPolygon/polygon-edge/contracts"
 	"github.com/0xPolygon/polygon-edge/txrelayer"
@@ -23,7 +23,7 @@ var params withdrawParams
 
 func GetCommand() *cobra.Command {
 	unstakeCmd := &cobra.Command{
-		Use:     "withdraw-child",
+		Use:     "withdraw",
 		Short:   "Withdraws pending withdrawals on child chain for given validator",
 		PreRunE: runPreRun,
 		RunE:    runCommand,
